+++ blueprints/ember-cli-sticky/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  description: 'Ember addon for Sticky.js',
-
-  normalizeEntityName: function() {}, // no-op since we're just adding dependencies  
-
-  afterInstall: function() {
-    return this.addBowerPackageToProject('sticky');
-  }
-};